pyproject.toml
[tool.poetry]
name = "ichiba_listing_violations_phase2_experiment"
version = "0.1.0"
description = "A template for general data science project"
authors = ["tianjia.a.ni <tianjia.a.ni@rakuten.com>"]
readme = "README.md"
packages = [
	{include = "ilv2", from = "src"},
]

[tool.poetry.dependencies]
python = "~3.9.16"
llvmlite = "^0.39.1"
mlflow = "^2.14.1"
scikit-learn = "^1.2.2"
numpy = "^1.24.2"
pandas = "^1.5.3"
polars = "^0.16.11"
snakeviz = "^2.1.1"
colorlog = "^6.7.0"
hydra-core = "^1.3.2"
black = "^23.1.0"
isort = "^5.12.0"
autopep8 = "^2.0.2"
torch = "2.0.0"
pytorch-lightning = "2.0.0"
torchvision = "^0.15.1"
rich = "^13.3.2"
google-cloud-secret-manager = "^2.16.4"
ipykernel = "^6.29.4"
ruamel-yaml = "^0.18.6"
seaborn = "^0.13.2"
plotly = "^5.22.0"
google-cloud-storage = "^2.16.0"
janome = "^0.5.0"
pandarallel = "^1.6.5"
japanize-matplotlib = "^1.1.3"
imbalanced-learn = "^0.12.3"
beautifulsoup4 = "^4.12.3"
openpyxl = "^3.1.5"
transformers = "^4.44.2"
sentencepiece = "^0.2.0"
bitsandbytes = "^0.39.0"
triton = "2.0.0"
nltk = "^3.9.1"
rouge = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"